updates store. sessions are now separated as upcomingrealtimesessions and fulfilledrequests
